+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- The compiler will now produce more useful error messages when two or more nodes in a compilation share the same name.
+
 ### Removed
 
 ## [2.2.3] 2022-08-28